pyproject.toml
[project]
name = "django-playground"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=5.2.8",
    "django-bootstrap5>=26.1",
    "django-cleanup>=9.0.0",
    "django-extensions>=4.1",
    "django-filter>=25.2",
    "djangorestframework>=3.16.1",
    "pillow>=12.1.0",
]

[dependency-groups]
check = [
    "djlint>=1.36.4",
]
dev = [
    "django-debug-toolbar>=6.1.0",
    "django-extensions>=4.1",
    "ipython>=9.7.0",
    "werkzeug>=3.1.5",
]

[tool.uv]
default-groups = ["dev","check"]

[tool.djlint]
profile = "django"
max_blank_lines = 1
line_break_after_multiline_tag = true
